osint-bot/requirements.txt
python-telegram-bot==13.15
requests
python-whois
dnspython
pyexiftool
redis
python-dotenv
phonenumbers